完成投稿接口的实现，修改了jwt的payload参数 (#11)
* 进行了项目结构的重构

* add 3 default videos

* modify video feed data

* 同步到远端

* 完成视频流随机推送的功能开发，为Video加入了Title项

* 完成投稿接口的实现，修改了jwt的payload参数

+++ model/video.go
@@ -6,7 +6,7 @@ type Video struct {
 	Title     string `gorm:"column:title"`
 	PlayUrl   string `gorm:"column:play_url"`
 	CoverUrl  string `gorm:"column:cover_url"`
-	CreatedAt int
+	CreatedAt int64
 	Comments  []*Comment
 	Likes     []*User `gorm:"many2many:like;joinForeignKey:video_id;joinReferences:user_id;"`
 }
